Update analyser to work with new format
This update affects all the analyser scripts, so that they work with the
new format of parsed constitutions as produced by the constitution
parser.

There are further changes with how the original constitution text is
rewritten to make section titles and references more consistent.

The Section class format is changed so that the information in section
objects is self-contained.